Backport to 4.2branch of: SVN commit 967638 by darioandres:
When updating the gamma values, first search for the "xorg.conf" file
which is the default config file in this moment. Then, fallback to
"XF86Config"
This solves an issue on some distributions as OpenSUSE which creates both
xorg.conf and XF86Config. And the gamma values were only applied to XF86Config
which is not used.

CCBUG: 192484


svn path=/branches/KDE/4.2/kdegraphics/kgamma/; revision=967639