Apply user setting at login/startup
KGamma's saved user settings are not applied on startup/login. The user
has to enter the KCM to apply them.
This makes it rather useless, as not even saving the settings
system-wide really works any more. (this requires an xorg.conf which
normally doesn't exist nowadays)

This patch uses kcminit to apply these settings again on login.
Apparently this has been forgotten to be adapted when moving/porting
kgamma to KDE4.

BUG: 218668
FIXED-IN: 4.13.3
REVIEW: 118898